requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
av==16.0.1
bcrypt==3.2.2
bert-score==0.3.13
blis==0.7.11
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.3
ctranslate2==4.6.1
cycler==0.12.1
cymem==2.0.13
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
faiss-cpu==1.13.0
fastapi==0.122.0
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.10.0
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6
nltk==3.9.2
numpy==1.26.4
ollama==0.6.1
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
preshed==3.0.12
protobuf==6.33.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
pypdfium2==5.1.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rouge_score==0.1.2
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.9.4
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.3.4
wrapt==2.0.1
